refactor(FE): useLabel to useSearch 훅 이름 변경 적용 (#355)
